// src/app/models/enums/meal-type-general-enum.ts
export enum MealTypeGeneralEnum {
  DANIE_GLOWNE = 'DANIE_GLOWNE',
  ZUPA = 'ZUPA',
  SUROWKA = 'SUROWKA',
  DODATKI = 'DODATKI',
  NAPOJ = 'NAPOJ',
  PIECZYWO = 'PIECZYWO',
  WEDLINA = 'WEDLINA',
  NABIAL = 'NABIAL',
  WARZYWO = 'WARZYWO',
  PODWIECZOREK = 'PODWIECZOREK',
  DRUGIE_SNIADANIE = 'DRUGIE_SNIADANIE'
}